feat: recurring delete donation modal (#100)
* feat: preliminary modal changes

* feat: connect frontend and backend for recurring donations

* refactor: linting

* fix: fix ability to delete each type of event

* refactor: linting and renaming params

* refactor: remove console.log statement

* fix: add recurring behaviour beyond a certain date

* refactor: linting and passing tests

* refactor: comments and unnecessary lines

* refactor: linting -.-

* chore: reword single vs multidonation

+++ backend/typescript/rest/schedulingRoutes.ts
@@ -120,16 +120,27 @@ schedulingRouter.put("/:id", updateSchedulingDtoValidator, async (req, res) => {
 });
 
 /* Delete scheduling by id (e.g. /scheduling/63)
-  and takes optional query for recurring donation 
-  id to delete by recurring donations (e.g. /scheduling/63?recurringDonationId=1)
+  or deletes by recurring donation id 
+  (e.g. /scheduling?recurringDonationId=1?currentDate=2022-01-31T05:00:00.000Z)
 */
 schedulingRouter.delete("/:id?", async (req, res) => {
   const { id } = req.params;
-  const { recurringDonationId } = req.query;
+  const { recurringDonationId, currentDate } = req.query;
+  const contentType = req.headers["content-type"];
+
+  if (id && recurringDonationId) {
+    await sendResponseByMimeType(res, 400, contentType, [
+      {
+        error: "Cannot delete by both id and recurringSchedulingId",
+      },
+    ]);
+    return;
+  }
   if (recurringDonationId) {
     try {
       await schedulingService.deleteSchedulingByRecurringDonationId(
         recurringDonationId as string,
+        currentDate as string,
       );
       res.status(204).send();
     } catch (error: unknown) {
